requirements.txt
fastapi
uvicorn
pydantic
langchain-groq
langchain-core
streamlit
requests
python-dotenv